# api.yaml
openapi: 3.0.3
info:
  title: ARCHIVE API
  version: 1.0.0
  description: |
    **ARCHIVE-BE** — 개인 생산성 및 회고 관리 백엔드 API.

    Clean Architecture 기반 (Domain → Application → Infrastructure → Presentation).

    ### 인증 방식
    - **Access Token**: `Authorization: Bearer <JWT>` 헤더 (만료 15분)
    - **Refresh Token**: `refresh_token` HttpOnly Cookie (만료 7일, `Secure`, `SameSite=Lax`)

    ### 공통 응답 포맷
    모든 응답은 아래 래퍼를 사용한다.
    ```json
    { "status": "success" | "accepted" | "error", "code": "OK", "data": { ... } }
    ```
    에러 응답은 `data: null`이며 `details[]`가 추가된다.

    ### AI 요약 비동기 처리
    `POST /summaries/generate`는 `202 Accepted`만 반환한다.
    실제 결과는 Celery worker가 처리하며, 클라이언트는 `GET /summaries/{id}/stream` SSE로 완료 신호를 받는다.
servers:
  - url: http://localhost:8000/api/v1
    description: 로컬 개발 서버
  - url: https://api.archive.example.com/api/v1
    description: 운영 서버 (placeholder)
...
security:
  - BearerAuth: []

# ────────────────────────────────────────────────────────────────────────────────
paths:
...
  /entries:
    get:
      tags: [entries]
      summary: 회고 엔트리 목록 조회
      description: 회고 종류·기간 필터로 조회한다.
      parameters:
        - in: query
          name: retroType
          schema: { $ref: '#/components/schemas/RetroType' }
        - in: query
          name: from
          schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
        - in: query
          name: to
          schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseEntryList' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'

    post:
      tags: [entries]
      summary: 회고 엔트리 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntryCreateRequest' }
            example:
              date_key: "2026-05-31"
              title: 5월 31일 회고
              content: 오늘은 ...
              retro_type: daily
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseEntry' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '409':
          $ref: '#/components/responses/Conflict_409'
        '422': { $ref: '#/components/responses/ValidationError_422' }
      x-error-codes:
        - JOURNAL_ENTRY_ALREADY_EXISTS
  /entries/{entry_id}:
    parameters:
      - $ref: '#/components/parameters/EntryIdPath'

    get:
      tags: [entries]
      summary: 단일 엔트리 조회
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseEntry' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '404':
          $ref: '#/components/responses/NotFound_404'
      x-error-codes:
        - JOURNAL_ENTRY_NOT_FOUND

    put:
      tags: [entries]
      summary: 엔트리 Upsert
      description: 존재하면 수정, 없으면 생성한다 (idempotent).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntryUpsertRequest' }
      responses:
        '200':
          description: Upsert 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseEntry' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '422': { $ref: '#/components/responses/ValidationError_422' }

    delete:
      tags: [entries]
      summary: 엔트리 삭제
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseEmpty' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '404':
          $ref: '#/components/responses/NotFound_404'
      x-error-codes:
        - JOURNAL_ENTRY_NOT_FOUND

  # ════════════════════════════════════════════════════════════════════════════
  # SUMMARIES (AI 요약)
  # ════════════════════════════════════════════════════════════════════════════
  /summaries/generate:
    post:
      tags: [summaries]
      summary: AI 요약 생성 요청 (비동기)
      description: |
        Celery worker(`ai_tasks` 큐, priority=9)로 요약 작업을 enqueue하고 **즉시 202를 반환**한다.

        - 같은 기간에 PENDING/IN_PROGRESS 요약이 있으면 `409 RETRO_SUMMARY_ALREADY_IN_PROGRESS` 반환
        - COMPLETED 요약이 있으면 그대로 반환 (재실행하지 않음)
        - FAILED 요약은 자동으로 PENDING 리셋 후 재실행
        - Gemini 호출은 rate_limit `60/min` 적용

        ### 데이터 소스 정책 (수동·자동 동일)
        - **weekly** : 그 주의 일일 회고(JournalEntry) 직접 사용
        - **monthly**: **하이브리드** — 주마다 weekly 요약 있으면 그것 사용,
          없으면 그 주 entries 첨부. weekly 요약이 있더라도 갱신 이후 추가된 entry 가
          있으면 보강(방식 B) — 데이터 손실 0.
        - **annual** : **2단계 하이브리드** — 월마다 monthly 요약 있으면 그것,
          없으면 그 달의 weekly 요약들로 보강. 둘 다 없으면 해당 월 스킵.

        ### 사전 점검 권장
        monthly/annual 생성 전에 `GET /summaries/readiness` 호출 권장 — 데이터가
        부족(`recommendation: insufficient`)하면 사용자에게 다이얼로그를 띄워
        확인 받기.
      parameters:
        - in: query
          name: type
          required: true
          schema: { $ref: '#/components/schemas/SummaryType' }
        - in: query
          name: periodStart
          schema: { type: string, format: date }
          description: |
            생략 시 직전 기간 자동 계산 (지난 주 월~일 / 지난 달 / 작년).
            지정 시 해당 기간의 시작일 (월요일/1일/1월1일).
      responses:
        '202':
          description: 요청 접수
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseSummary' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '409':
          $ref: '#/components/responses/Conflict_409'
      x-error-codes:
        - RETRO_SUMMARY_ALREADY_IN_PROGRESS
        - RETRO_SUMMARY_INVALID_STATE
  /summaries/readiness:
    get:
      tags: [summaries]
      summary: 요약 생성 사전 점검 (entry 밀도 기반)
      description: |
        monthly/annual 요약을 생성하기 전에 사용자의 회고 **원천 데이터 밀도**를
        측정해 FE 가 "데이터 부족" 다이얼로그를 띄울지 결정하게 한다.

        - **측정 지표**: child summary 존재 여부 아님 — **entry 밀도** 기반
          - monthly: `expectedUnits` = 그 달의 일수, `coveredUnits` = entry 가 있는 unique 날짜 수
          - annual : `expectedUnits` = 12, `coveredUnits` = entry 가 있는 월 수
        - **임계값**: `completenessRatio < 0.7` 이면 `recommendation: "insufficient"`,
          이상이면 `"ok"`.
        - **weekly 는 지원 안 함** — 호출 시 `422 RETRO_SUMMARY_READINESS_UNSUPPORTED`

        ### 권장 FE 흐름
        1. 사용자가 [회고 생성] 클릭 → 이 API 호출
        2. `recommendation == "insufficient"` → 다이얼로그 띄우기
           ("entry 가 X일/30일 작성됨. 그대로 진행하시겠습니까?")
        3. 사용자가 "진행" → `POST /summaries/generate`
      parameters:
        - in: query
          name: type
          required: true
          schema:
            type: string
            enum: [monthly, annual]
          description: weekly 는 422 반환.
        - in: query
          name: periodStart
          schema: { type: string, format: date }
          description: 생략 시 직전 기간(지난 달 / 작년).
      responses:
        '200':
          description: 점검 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseSummaryReadiness' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '422':
          $ref: '#/components/responses/ValidationError_422'
      x-error-codes:
        - RETRO_SUMMARY_READINESS_UNSUPPORTED
  /summaries:
    get:
      tags: [summaries]
      summary: 요약 목록 조회 (타입별)
      parameters:
        - in: query
          name: type
          required: true
          schema: { $ref: '#/components/schemas/SummaryType' }
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseSummaryList' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
  /summaries/{summary_id}:
    parameters:
      - $ref: '#/components/parameters/SummaryIdPath'

    get:
      tags: [summaries]
      summary: 단일 요약 조회
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseSummary' }
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '404':
          $ref: '#/components/responses/NotFound_404'
      x-error-codes:
        - RETRO_SUMMARY_NOT_FOUND
  /summaries/{summary_id}/stream:
    parameters:
      - $ref: '#/components/parameters/SummaryIdPath'

    get:
      tags: [summaries]
      summary: 요약 완료 SSE 스트림
      description: |
        Celery worker가 요약 처리를 완료할 때까지 대기하며, 완료 시 단일 SSE 이벤트를 전송하고 종료한다.

        - **타임아웃**: 5분 (이후 `{"status": "timeout"}` 전송)
        - 이미 완료/실패 상태면 즉시 이벤트 전송 후 종료

        ### 이벤트 페이로드
        ```
        data: {"status": "completed", "summary_id": "summ_..."}
        data: {"status": "failed",    "summary_id": "summ_..."}
        data: {"status": "timeout",   "summary_id": "summ_..."}
        data: {"status": "error",     "error": "<ERROR_CODE>"}
        ```
      responses:
        '200':
          description: SSE 스트림
          content:
            text/event-stream:
              schema:
                type: string
                example: 'data: {"status":"completed","summary_id":"summ_abc"}\n\n'
        '401':
          $ref: '#/components/responses/Unauthorized_401'

  # ════════════════════════════════════════════════════════════════════════════
  # NOTIFICATIONS
  # ════════════════════════════════════════════════════════════════════════════
...
components:
...
  securitySchemes:
  # ════════════════════════════════════════════════════════════════════════════
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '`Authorization: Bearer <access_token>` (만료 15분)'
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly · Secure · SameSite=Lax · Max-Age 7d
    OnboardingCookie:
      type: apiKey
      in: cookie
      name: onboarding_token
      description: |
        OAuth 신규 사용자 온보딩용 임시 토큰. HttpOnly · Secure · SameSite=Lax · Max-Age 30분.
        Redis에 `auth:onboarding:{token} → {provider, provider_user_id, email}` 형태로 적재됨.
        클라이언트가 `POST /auth/oauth/onboarding`에서 사용 후 즉시 삭제된다.

  # ════════════════════════════════════════════════════════════════════════════
  parameters:
...
    EntryIdPath:
      in: path
      name: entry_id
      required: true
      schema: { type: string }
      example: entry_01HXYZ...
    SummaryIdPath:
      in: path
      name: summary_id
      required: true
      schema: { type: string }
      example: summ_01HXYZ...
...
  responses:
...
    Unauthorized_401:
      description: |
        인증 실패. 가능한 코드:
        - `AUTH_TOKEN_EXPIRED` — access token 만료 (**FE 자동 refresh 트리거**)
        - `AUTH_TOKEN_INVALID` — 토큰 누락 / 형식 오류 / 서명 불일치 / type mismatch
        - `AUTH_INVALID_CREDENTIALS` — 이메일/비밀번호 불일치
        - `AUTH_REFRESH_TOKEN_INVALID` / `AUTH_REFRESH_TOKEN_REVOKED` — refresh 실패
        - `AUTH_REFRESH_TOKEN_REUSE_DETECTED` — 폐기된 refresh token 재사용 감지(탈취 의심). 해당 사용자의 모든 세션이 즉시 폐기됨. 사용자에게 보안 알림 권장 후 재로그인 유도.
        - `AUTH_OAUTH_STATE_INVALID` — OAuth state 검증 실패
        - `AUTH_ONBOARDING_TOKEN_INVALID` — onboarding cookie 누락
        - `AUTH_ONBOARDING_TOKEN_EXPIRED` — onboarding TTL(30분) 만료
        - `AUTH_PASSWORD_RESET_TOKEN_INVALID` — 토큰 누락/형식 오류
        - `AUTH_PASSWORD_RESET_TOKEN_EXPIRED` — 토큰 TTL(30분) 만료 또는 이미 사용됨
        - `GITHUB_TOKEN_INVALID` — GitHub OAuth 토큰 만료/폐기 (FE는 GitHub OAuth 재인증 유도)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: AUTH_TOKEN_EXPIRED
            data: null
            details: []
    NotFound_404:
      description: |
        리소스를 찾을 수 없음. 가능한 코드:
        - 도메인별 NOT_FOUND (`TODO_NOT_FOUND`, `JOURNAL_ENTRY_NOT_FOUND`, `RETRO_SUMMARY_NOT_FOUND`, `NOTIFICATION_NOT_FOUND`, `USER_NOT_FOUND`, `GITHUB_REPOSITORY_NOT_FOUND`)
        - `AUTH_SESSION_NOT_FOUND` — 폐기 대상 sessionId 가 본인 소유가 아니거나 이미 만료됨
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: TODO_NOT_FOUND
            data: null
            details: []
    Conflict_409:
      description: |
        상태 충돌 (중복·이미 진행 중 등). 가능한 코드:
        - `USER_EMAIL_DUPLICATED` — 이메일 중복 가입
        - `TODO_ALREADY_COMPLETED` / `TODO_ALREADY_IN_PROGRESS`
        - `JOURNAL_ENTRY_ALREADY_EXISTS`
        - `RETRO_SUMMARY_ALREADY_IN_PROGRESS`
        - `GITHUB_REPOSITORY_ALREADY_LINKED`
        - `AUTH_OAUTH_ACCOUNT_ALREADY_LINKED` — link 시도한 OAuth 계정이 다른 사용자에 이미 연결됨
        - `AUTH_OAUTH_PROVIDER_ALREADY_LINKED` — 현재 사용자가 같은 provider에 다른 계정으로 이미 연결됨
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: USER_EMAIL_DUPLICATED
            data: null
            details: []
    ValidationError_422:
      description: |
        요청 검증 실패. 두 가지 경로:
        - **Pydantic 스키마 에러** → `code: VALIDATION_ERROR`, `details` 에 field별 메시지
        - **도메인 검증 에러** (핸들러가 직접 반환) → 도메인 코드 사용:
          - `AUTH_COUNTRY_INVALID` — ISO 3166-1 미등록 국가, 또는 tz 데이터 없는 국가
          - `AUTH_COUNTRY_TIMEZONE_REQUIRED` — 다중 tz 국가에서 timezone 누락
          - `AUTH_TIMEZONE_INVALID` — IANA 형식 아니거나 해당 국가의 옵션 밖
          - `RETRO_SUMMARY_READINESS_UNSUPPORTED` — readiness 점검은 monthly/annual 만 지원, weekly 는 거부
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            status: error
            code: VALIDATION_ERROR
            data: null
            details:
              - field: email
                message: value is not a valid email address
...
  schemas:
...
    ApiResponseEmpty:
      type: object
      required: [status, code, data]
      properties:
        status: { type: string, enum: [success, accepted] }
        code:   { type: string, example: OK }
        data:   { type: 'null' }
    ErrorResponse:
      type: object
      required: [status, code, data]
      properties:
        status: { type: string, enum: [error] }
        code:
          type: string
          description: 도메인 에러 코드 (예 `AUTH_TOKEN_EXPIRED`, `TODO_NOT_FOUND`, `VALIDATION_ERROR`)
        data: { type: 'null' }
        details:
          type: array
          items:
            type: object
            properties:
              field:   { type: string }
              message: { type: string }

    # ── Enum 정의 ───────────────────────────────────────────────────────────
...
    RetroType:
      type: string
      enum: [daily, weekly, monthly, yearly]
      description: 회고 엔트리 종류
    SummaryType:
      type: string
      enum: [weekly, monthly, annual]
      description: AI 요약 주기
...
    SummaryStatus:
      type: string
      enum: [pending, in_progress, completed, failed]
...
    EntryCreateRequest:
      type: object
      required: [date_key, title]
      properties:
        date_key:   { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
        title:      { type: string }
        content:    { type: string, default: "" }
        retro_type:
          $ref: '#/components/schemas/RetroType'
          default: daily
    EntryUpsertRequest:
      $ref: '#/components/schemas/EntryCreateRequest'
    GithubPushResponse:
      type: object
      required: [pushedAt, commitSha, htmlUrl, path, repositoryFullName]
      description: |
        해당 회고록이 GitHub 에 push 되어 있다는 신호. push 단위는 (period_type, period_key)
        — 같은 period 에 entry/summary 가 같이 있으면 둘 다 동일한 push 레코드를 가리킨다.
        (GitHub 측 파일이 단일이므로 의도된 동작.)
      properties:
        pushedAt:           { type: string, format: date-time }
        commitSha:          { type: string }
        htmlUrl:            { type: string, format: uri }
        path:
          type: string
          description: 저장소 내 파일 경로 (예 'daily/2026-06-08.md')
        repositoryFullName: { type: string, example: 'owner/archive' }
    EntryResponse:
      type: object
      required: [id, user_id, date_key, title, content, retro_type, created_at]
      properties:
        id:         { type: string }
        user_id:    { type: string }
        date_key:   { type: string }
        title:      { type: string }
        content:    { type: string }
        retro_type: { $ref: '#/components/schemas/RetroType' }
        githubPush:
          allOf:
            - $ref: '#/components/schemas/GithubPushResponse'
          nullable: true
          description: 해당 (retro_type, date_key) 기반 period 에 push 된 적이 없으면 null
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time, nullable: true }
    ApiResponseEntry:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/EntryResponse' }
    ApiResponseEntryList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/EntryResponse' }

    # ── Summary ────────────────────────────────────────────────────────────
    SummaryContentResponse:
      type: object
      properties:
        achievements: { type: array, items: { type: string } }
        challenges:   { type: array, items: { type: string } }
        learnings:    { type: array, items: { type: string } }
        next_focus:   { type: array, items: { type: string } }
    SummaryResponse:
      type: object
      required: [id, user_id, summary_type, period_start, period_end, status, created_at]
      properties:
        id:           { type: string }
        user_id:      { type: string }
        summary_type: { $ref: '#/components/schemas/SummaryType' }
        period_start: { type: string, format: date }
        period_end:   { type: string, format: date }
        status:       { $ref: '#/components/schemas/SummaryStatus' }
        content:
          $ref: '#/components/schemas/SummaryContentResponse'
          nullable: true
        githubPush:
          allOf:
            - $ref: '#/components/schemas/GithubPushResponse'
          nullable: true
          description: 해당 summary 의 (summary_type, period_start) 기반 period 에 push 된 적이 없으면 null
        created_at:   { type: string, format: date-time }
        updated_at:   { type: string, format: date-time, nullable: true }
    ApiResponseSummary:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/SummaryResponse' }
    ApiResponseSummaryList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/SummaryResponse' }
    SummaryReadinessResponse:
      type: object
      required:
        - summaryType
        - periodStart
        - periodEnd
        - expectedUnits
        - coveredUnits
        - entryCount
        - completenessRatio
        - recommendation
      properties:
        summaryType:
          type: string
          enum: [monthly, annual]
        periodStart:        { type: string, format: date }
        periodEnd:          { type: string, format: date }
        expectedUnits:
          type: integer
          description: monthly = 그 달의 일수, annual = 12
        coveredUnits:
          type: integer
          description: monthly = entry 있는 unique 날짜 수, annual = entry 있는 월 수
        entryCount:
          type: integer
          description: 기간 내 총 entry 수 (참고용, 임계값에 직접 영향 없음)
        completenessRatio:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: coveredUnits / expectedUnits, 소수 4자리 반올림
        recommendation:
          type: string
          enum: [ok, insufficient]
          description: completenessRatio >= 0.7 이면 ok, 미만이면 insufficient
    ApiResponseSummaryReadiness:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEmpty'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/SummaryReadinessResponse' }

    # ── Notification ───────────────────────────────────────────────────────
...